+++ potato-plugins/potato-plugin-hbase/src/main/scala/spark/streaming/potato/plugins/hbase/sink/HBaseSinkUtil.scala
@@ -7,7 +7,7 @@ import org.apache.spark.streaming.dstream.DStream
 import spark.streaming.potato.plugins.hbase.SerializableConfiguration
 import spark.streaming.potato.plugins.hbase.TableUtil._
 
-object HBaseSink extends Logging {
+object HBaseSinkUtil extends Logging {
   def saveToHBase(rdd: RDD[MutationAction], conf: SerializableConfiguration, table: String, bufferSize: Int): Unit = {
     rdd.foreachPartition { part =>
       withMutator(conf, table) { mutator =>
@@ -35,29 +35,6 @@ object HBaseSink extends Logging {
     }
   }
 
-}
-
-object HBaseSinkImplicits {
-
-  class MutationActionRDD(rdd: RDD[MutationAction]) extends Serializable {
-    def saveToHBase(conf: SerializableConfiguration, table: String, bufferSize: Int = 1000): Unit = {
-      HBaseSink.saveToHBase(rdd, conf, table, bufferSize)
-    }
-  }
-
-  class MutationActionDStream(stream: DStream[MutationAction]) extends Serializable {
-    def saveToHBase(conf: SerializableConfiguration, table: String, bufferSize: Int = 1000): Unit = {
-      HBaseSink.saveToHBase(stream, conf, table, bufferSize)
-    }
-  }
-
-  implicit def toMutationActionRDD(rdd: RDD[MutationAction]): MutationActionRDD = {
-    new MutationActionRDD(rdd)
-  }
-
-  implicit def toMutationActionDStream(stream: DStream[MutationAction]): MutationActionDStream = {
-    new MutationActionDStream(stream)
-  }
 }
 
 case class MutationAction(action: MutationType.Type, mutation: Mutation)
