+++ potato-quickstart/src/main/scala/spark/streaming/potato/quickstart/Kafka2HBaseDemo.scala
@@ -6,15 +6,14 @@ import org.apache.hadoop.hbase.client.Put
 import org.apache.hadoop.hbase.util.Bytes
 import org.apache.spark.streaming.StreamingContext
 import org.apache.spark.streaming.dstream.DStream
-import spark.streaming.potato.plugins.hbase.sink.HBaseSinkImplicits._
-import spark.streaming.potato.plugins.hbase.sink.{MutationAction, MutationType}
-import spark.streaming.potato.plugins.kafka.source.KafkaSource
+import spark.streaming.potato.plugins.hbase.sink._
+import spark.streaming.potato.plugins.kafka.source.KafkaSourceUtil
 import spark.streaming.potato.plugins.kafka.source.offsets.OffsetsManager
 import spark.streaming.potato.template.template.KafkaSourceTemplate
 
 object Kafka2HBaseDemo extends KafkaSourceTemplate[String] {
   override def initKafka(ssc: StreamingContext): (DStream[String], OffsetsManager) =
-    KafkaSource.valueDStream(ssc)
+    KafkaSourceUtil.valueDStream(ssc)
 
   override def doWork(args: Array[String]): Unit = {
     getStream.transform { rdd =>
