+++ potato-quickstart/src/main/scala/spark/streaming/potato/quickstart/Kafka2KafkaDemo.scala
@@ -3,14 +3,14 @@ package spark.streaming.potato.quickstart
 import org.apache.kafka.clients.producer.ProducerRecord
 import org.apache.spark.streaming.StreamingContext
 import org.apache.spark.streaming.dstream.DStream
-import spark.streaming.potato.plugins.kafka.source.KafkaSource
+import spark.streaming.potato.plugins.kafka.source.KafkaSourceUtil
 import spark.streaming.potato.plugins.kafka.source.offsets.OffsetsManager
 import spark.streaming.potato.template.template.KafkaSourceTemplate
-import spark.streaming.potato.plugins.kafka.sink.KafkaSinkImplicits._
+import spark.streaming.potato.plugins.kafka.sink._
 
 object Kafka2KafkaDemo extends KafkaSourceTemplate[(String, String)] {
   override def initKafka(ssc: StreamingContext): (DStream[(String, String)], OffsetsManager) =
-    KafkaSource.kvDStream(ssc)
+    KafkaSourceUtil.kvDStream(ssc)
 
   override def doWork(args: Array[String]): Unit = {
     getStream.map { f =>
