+++ potato-template/src/test/scala/spark/streaming/potato/template/template/KafkaErrTest.scala
@@ -4,7 +4,7 @@ import org.apache.spark.SparkConf
 import org.apache.spark.internal.Logging
 import org.apache.spark.streaming.StreamingContext
 import org.apache.spark.streaming.dstream.DStream
-import spark.streaming.potato.plugins.kafka.source.KafkaSource
+import spark.streaming.potato.plugins.kafka.source.KafkaSourceUtil
 import spark.streaming.potato.common.conf.CommonConfigKeys._
 import spark.streaming.potato.plugins.kafka.KafkaConfigKeys._
 import spark.streaming.potato.plugins.kafka.source.offsets.OffsetsManager
@@ -12,7 +12,7 @@ import spark.streaming.potato.plugins.lock.LockConfigKeys._
 
 object KafkaErrTest extends KafkaSourceTemplate[(String, String)] with Logging {
   override def initKafka(ssc: StreamingContext): (DStream[(String, String)], OffsetsManager) =
-    KafkaSource.kvDStream(ssc)
+    KafkaSourceUtil.kvDStream(ssc)
 
   override def doWork(args: Array[String]): Unit = {
     getStream.foreachRDD { rdd =>
