+++ potato-template/src/test/scala/spark/streaming/potato/template/template/KafkaTopn1Test.scala
@@ -7,12 +7,12 @@ import org.apache.spark.streaming.dstream.DStream
 import spark.streaming.potato.common.conf.CommonConfigKeys._
 import spark.streaming.potato.plugins.kafka.KafkaConfigKeys._
 import spark.streaming.potato.plugins.kafka.source.offsets.OffsetsManager
-import spark.streaming.potato.plugins.kafka.source.KafkaSource
+import spark.streaming.potato.plugins.kafka.source.KafkaSourceUtil
 import spark.streaming.potato.plugins.lock.LockConfigKeys._
 
 object KafkaTopn1Test extends KafkaSourceTemplate[(String, String)] with Logging {
   override def initKafka(ssc: StreamingContext): (DStream[(String, String)], OffsetsManager) =
-    KafkaSource.kvDStream(ssc)
+    KafkaSourceUtil.kvDStream(ssc)
 
   override def doWork(args: Array[String]): Unit = {
     getStream.foreachRDD { rdd =>
